// tests/enums/data-for-bodies.ts
export enum pageTitles {
  integrationTests = 'integrationTests',
  defaultEngTitle = 'Default English Title',
  stableEngTitle = 'Stable English Title',
  unpublishPageTitle = 'Unpublish Page Test',
  draftPageTitle = 'Test Draft Page',
  draftBannerTitle = 'Test Banner Page',
  updatedTitle = 'updated title',
  pageToDelete = 'Page To Delete',
  scheduledPageTitle = 'Scheduled Blog Page',
  pageWithImage = 'Page with image',
  koreanTitle = '현지화된 블로그',
  chineseTitle = '本地化博客',
  seoPage = 'SEO page'
}

export enum locales {
  defaultLocale = 'en',
  koreanCMSLocale = 'ko',
  chineseCMSLocale = 'zh-CN',
  koreanFELocale = 'kr',
  chineseFELocale = 'cn'
}

export enum pageDescriptions {
  defaultDescription = 'This is a test blog description.',
  defaultKoreanDescription = '업데이트되었습니다.',
  defaultChineseDescription = '元描述',
  initialMetaDescription = 'Initial meta description.',
  updatedMetaDescription = 'Updated meta description.',
  updatedKoreanMetaDescription = '메타 설명이 업데이트되었습니다.',
  updatedChineseMetaDescription = '更新了元描述。'
}

export enum promoBannerTitles {
  defaultPromoTitle = 'Promo',
  defaultBannerTitle = 'Default Banner Title',
  testBannerTitle = 'Test Banner Title',
}

export enum promoDescriptions {
  defaultTitle = 'Promo Title',
  defaultDescription = 'This is a test promo description.',
}

export enum promoBodyData {
  descriptionForMechanics = '7.All marketing prizes to be shouldered by Solaire Resort Entertainment City.' +
    '8.     This offer is not valid in conjunction with other promotions.',
  descriptionForTerms =  'd.     By participating in the promotion, the member is deemed to have accepted and agreed to be bound by these Terms and conditions that Solaire Resort Entertainment City may issue from time to time.\n' +
  'e.      The management reserves the right to withhold and cancel the awarding of prizes in case of any fraud, but in consultation with PAGCOR.\n' +
  'f.       Solaire Resort Entertainment City reserves the right to modify, change, and revise the rules, prizes, and mechanics of this promotion without prior notice, but in consultation with PAGCOR.\n' +
  'g.      Solaire Resort Entertainment City shall not be liable to any person in the program for any injuries, losses, or damages in respect of, in connection with and/ or arising from the program.\n' +
  'h.     All matters and disputes arising out of this promotion will be subject to the final decision of Solaire Resort Entertainment City.\n' +
  'i.       Under no circumstances shall Solaire Resort Entertainment City be liable for any discrepancies or mistakes which may occur due to errors in the system, computer malfunction or operational variances.\n' +
  'j.       In case of disputes, the Management’s decision is final.\n' +
  'k.      The Mobile App is a service of Bloomberry Resorts and Hotels, Inc. (“BRHI”). All promotions and offers at the Solaire Resort and Casino offered through this App are owned by BRHI, to the exclusion of Apple Inc. and Google Inc.'
}

export enum notificationMessages {
  defaultMessage = 'new notification',
  updatedMessage = 'updated notification message',
}

export enum notificationInfoTypes {
  warning = 'Warning',
  success = 'Success',
  error = 'Error',
}

export enum articlesCategory {
  sports = 'sports',
  integrated = 'integrated',
  press = 'press',
  events = 'events'
}

export enum invalidNames {
  invalidShortName = 'qw',
  invalid57Name = 'SuprelongnameSuprelongnameSuprelongnameSuprelongnameSupre',
  invalid256Name = 'SuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnamSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongnameSuprelongn',
  invalidOnlyDigitsName = '12345',
  invalidWithSpecialSymbolsName = 'Title@With#Symbols!'
}